+++ com.onesignal.unity.ios/Editor/BuildPostProcessor.cs
@@ -75,12 +75,12 @@ namespace OneSignalSDK {
         private static readonly string PluginLibrariesPath = Path.Combine(PackageName, "Runtime", "Plugins", "iOS");
         private static readonly string PluginFilesPath = Path.Combine("Packages", PluginLibrariesPath);
 
-        private readonly string _appGroupName = $"group.{PlayerSettings.applicationIdentifier}.onesignal";
-        
         private string _outputPath;
         private string _projectPath;
-        
+
         private readonly PBXProject _project = new PBXProject();
+
+        private string AppGroupName => $"group.{PlayerSettings.applicationIdentifier}.onesignal";
 
         /// <summary>
         /// must be between 40 and 50 to ensure that it's not overriden by Podfile generation (40) and that it's
@@ -153,7 +153,7 @@ namespace OneSignalSDK {
 
             projCapability.AddBackgroundModes(BackgroundModesOptions.RemoteNotifications);
             projCapability.AddPushNotifications(false);
-            projCapability.AddAppGroups(new[] { _appGroupName });
+            projCapability.AddAppGroups(new[] { AppGroupName });
             
             projCapability.WriteToFile();
         }
@@ -197,7 +197,7 @@ namespace OneSignalSDK {
             var entitlementsPath = GetEntitlementsPath(extensionGuid, ServiceExtensionTargetName);
             var projCapability = new ProjectCapabilityManager(_projectPath, entitlementsPath, ServiceExtensionTargetName);
             
-            projCapability.AddAppGroups(new[] { _appGroupName });
+            projCapability.AddAppGroups(new[] { AppGroupName });
             
             projCapability.WriteToFile();
 #endif